Fixed so that one can click on the rotator map and get it to rotate to that direction.

+++ trunk/c/openASC_Controller/src/mapviewwidget.h
@@ -8,6 +8,8 @@
 #include <QHBoxLayout>
 #include <QLabel>
 
+#include <QMouseEvent>
+
 //2 degrees target beamwidth
 #define TARGET_DIR_BEAMWIDTH	1
 #define TARGET_DIR_BEAMWIDTH_COLOR	red
@@ -21,16 +23,16 @@ public:
 	MapViewWidget( QWidget * parent = 0, Qt::WFlags f = 0 );
 	QPainter painter;
 	void setImagePath(QString path);
-	void setCurrentDir(float azimuthAngle, float beamWidth);
-	void setTargetDir(float azimuthAngle);
-	void setWidgetSize(float width, float height);
+	void setCurrentDir(int azimuthAngle, int beamWidth);
+	void setTargetDir(int azimuthAngle);
+	void setWidgetSize(int width, int height);
 	void setMapTitle(QString title);
 protected:
 		void paintEvent(QPaintEvent *event);
 		QString imagePath;
-		float currAzimuthAngle;
-		float currBeamWidth;
-		float targetAzimuthAngle;
+		int currAzimuthAngle;
+		int currBeamWidth;
+		int targetAzimuthAngle;
 		int sizeWidth;
 		int sizeHeight;
 		QString mapTitle;
